refactor of main script for segmentation
