+++ src/rigl_torch/utils/wandb_utils.py
@@ -1,4 +1,6 @@
-from typing import Callable, Any
+from typing import Callable, Any, Dict
+import omegaconf
+import wandb
 
 from rigl_torch.exceptions import WandbRunNameException
 
@@ -19,7 +21,7 @@ class WandbRunName:
             )
 
 
-def wandb_log_check(fn: Callable, log_to_wandb: bool = True) -> Callable:
+def _wandb_log_check(fn: Callable, log_to_wandb: bool = True) -> Callable:
     def wrapper(*args, **kwargs) -> Any:
         if log_to_wandb:
             return fn(*args, **kwargs)
@@ -27,3 +29,28 @@ def wandb_log_check(fn: Callable, log_to_wandb: bool = True) -> Callable:
             return None
 
     return wrapper
+
+
+def init_wandb(cfg: omegaconf.DictConfig, wandb_init_kwargs: Dict[str, Any]):
+    # We override logging functions now to avoid any calls
+    if not cfg.wandb.log_to_wandb:
+        print("No logging to WANDB! See cfg.wandb.log_to_wandb")
+        wandb.log = _wandb_log_check(wandb.log, cfg.wandb.log_to_wandb)
+        wandb.log_artifact = _wandb_log_check(
+            wandb.log_artifact, cfg.wandb.log_to_wandb
+        )
+        wandb.watch = _wandb_log_check(wandb.watch, cfg.wandb.log_to_wandb)
+        return None
+    _ = WandbRunName(name=cfg.experiment.name)  # Verify name is OK
+    run = wandb.init(
+        name=cfg.experiment.name,
+        entity=cfg.wandb.entity,
+        project=cfg.wandb.project,
+        config=omegaconf.OmegaConf.to_container(
+            cfg=cfg, resolve=True, throw_on_missing=True
+        ),
+        settings=wandb.Settings(start_method=cfg.wandb.start_method),
+        dir=cfg.paths.logs,
+        **wandb_init_kwargs,
+    )
+    return run
